% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{clshift}
\alias{clshift}
\title{Number of Crossings and Longest Run}
\usage{
clshift(seri, shift = 0, type = 0)
}
\arguments{
\item{seri}{numeric; seri a sequence of random draws}

\item{shift}{numeric; shift for the observatoobs}

\item{type}{numeric; 0 number of crossings, 1 longest run}
}
\value{
number of crossings or longest run, numeric
}
\description{
Auxiliary function for simclbin, computing
the number of crossings (type=0) or longest run (type=2)
in a sequence of independent normal observations. Crossings
and runs are related to whether the observations are above
a shift.
}
