% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossrunbin.r
\name{crossrunshift}
\alias{crossrunshift}
\title{wrapper for crossrunbin, success probability=pnorm(shift).}
\usage{
crossrunshift(nmax = 100, shift = 0, mult = 2, prec = 120, printn = FALSE)
}
\arguments{
\item{nmax}{max sequence length.}

\item{shift}{mean of normal distribution.}

\item{mult}{multiplier for joint probabilities.}

\item{prec}{mpft precision.}

\item{printn}{logical for progress output.}
}
\value{
list pt of joint probabilities. Cumulative probabilities
qt within each row are also included. Further, mostly for code
checking, lists pat and qat conditional on starting with a success,
and pbt and qbt conditional of starting with a failure, are
included.
}
\description{
wrapper for crossrunbin, success probability=pnorm(shift).
}
\examples{
crs15 <- crossrunshift(nmax=15,printn=TRUE)
print(crs15$pt[[15]])
}
