% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint100symm}
\alias{joint100symm}
\title{Joint probabilities, n=100, symmetric case}
\format{matrix, 100 rows and 100 columns}
\source{
generated by the function crossrunsymm and
transformed from an Rmpfr array to a matrix
}
\usage{
joint100symm
}
\description{
The joint probabilities of the number C og crossings
(0, ... 99) and the longest run L (1, ..., 100) in a
series of n=100 independent Bernoulli observations for the
symmetric case (success probability 0.5). The probabilities
are stored in the "times" representations, multiplied by
\eqn{2^{100-1}} and are integers in the symmetric
case. Only the joint distributions for n=15, 60, 100
and success probabilities 0.5 and 0.6 are included in
the package to avoid excessive storage, but many more
cases are generated in the script crossrun1.R.
}
\keyword{datasets}
