% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint14em}
\alias{joint14em}
\title{Joint probabilities, n=14, around the empirical median}
\format{
matrix, 13 rows and 7 columns
}
\source{
generated by the function crossrunsymm and
transformed from an Rmpfr array to a matrix
}
\usage{
joint14em
}
\description{
Joint probabilities of the number C of crossings
(1, ... 13) and the longest run L (1, ..., 17) in a
series of n=60 Bernoulli observations around its
empirical median. The probabilities  are stored in
the "times" representations, multiplied by
(60 by 30)/2, the number of constellations starting
above the median, and are integers. About the empirical
median there is at least one crossing, and the longest
run cannot exceed 14/2=7. Only the joint distributions
for n=14, 60 are included in the package to avoid excessive
storage, but many more cases may be generated by the function
'crossrunem. Since these computations are demanding in terms
of storage and computation time, they are at present not
performed for n much above 60.
}
\keyword{datasets}
