% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint14symm}
\alias{joint14symm}
\title{Joint probabilities, n=14, symmetric case}
\format{
matrix, 14 rows and 14 columns
}
\source{
generated by the function crossrunsymm and
transformed from an Rmpfr array to a matrix
}
\usage{
joint14symm
}
\description{
Joint probabilities of the number C of crossings
(0, ... 13) and the longest run L (1, ..., 14) in a
series of n=14 independent Bernoulli observations for the
symmetric case (success probability 0.5). The probabilities
are stored in the "times" representations, multiplied by
\eqn{2^{14-1}=8192} and are integers in the symmetric
case. Only the joint distributions for n=14, 60, 100
and success probabilities 0.5 and 0.6 are included in
the package to avoid excessive storage, but many more
cases may be generated by the function crossrunsymm.
}
\keyword{datasets}
