% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint60.6}
\alias{joint60.6}
\title{Joint probabilities, 60, success probability 0.6}
\format{matrix, 60 rows and 60 columns}
\source{
generated by the function crossrunbin and
transformed from an Rmpfr array to a matrix
}
\usage{
joint60.6
}
\description{
The joint probabilities of the number C og crossings
(0, ... 59) and the longest run L (1, ..., 60) in a
series of n=60 independent Bernoulli observations for
success probability 0.6. The probabilities are stored
in the "times" representations, multiplied by
\eqn{2^{60-1}}. Only the joint distributions for
n=15, 60, 100 and success probabilities 0.5 and 0.6 are
included in the package to avoid excessive storage, but
many more cases are generated in the script crossrun1.R.
}
\keyword{datasets}
