% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossrunem.r
\name{simclem}
\alias{simclem}
\title{Check of joint probabilities by simulations}
\usage{
simclem(m1 = 7, nsim = 1e+05)
}
\arguments{
\item{m1, }{half the sequence length}

\item{nsim}{number of simulations}
}
\value{
data frame with cs, number of crossings
and ls, longest run in the simulations.
}
\description{
Simulation of a sequence of n=2m observations
around the median in the sequence. To be used for checking
the results of crossrunem.
}
\examples{
simclem14 <- simclem(nsim=sum(joint14em))
print(table(simclem14)) # joint distributions in the simulations
print(joint14em) # for comparison

}
